Feat: explore 경로 prefetch 적용
